Cleanup Dashboard controller and routing.
I decided to use a not-Redis based routing system for now because it's far simpler and efficiency is not an issue now, readability is.
The Redis routing system was really complicated to reproduce in testing at the current stage (it is initialized at runtime) and, to me, it seems a premature over optimization.

The idea of using a shared base controller is good. I renamed it from SiteController (very generic ) to SubdominedController and flagged it as abstract.
For some reason other controllers, such as GemsController, don't use it while they should.

The DashboardController tests now exist and they work. I spent 3 hours fighting with Devise internals. Too much stuff there.